AddGradeView更新 在输入分数时可以通过Slider上新增的数字加Button方式以1分为step输入 增加了ScoreColor函数 以后在根据分数选择颜色时会用这个函数 依然没整理HomeView以外的代码结构

+++ StudyPulse/Managers/SubjectInfo.swift
@@ -9,25 +9,25 @@ import Foundation
 import Combine
 
 class SubjectInfo: ObservableObject {
-    func getMaxScore (level: String, subject: String) -> Int {
+    func getMaxScore (level: String, subject: String) -> Double {
         if level == "Middle School" {
             if subject == "Chinese" || subject == "Mathematics" || subject == "English" {
-                return 120
+                return 120.0
             } else if subject == "Science" {
-                return 160
+                return 160.0
             } else {
-                return 100
+                return 100.0
             }
         } else if level == "High School" {
             if subject == "Chinese" || subject == "Mathematics" || subject == "English" {
-                return 150
+                return 150.0
             } else {
-                return 100
+                return 100.0
             }
         } else if level == "Primary School" {
-            return 100
+            return 100.0
         }
-        return 100  
+        return 100.0
     }
 }
